+++ apps/backend/src/modules/projects/project.controller.ts
@@ -61,6 +61,7 @@ class ProjectController extends BaseController {
 				),
 			method: "POST",
 			path: ProjectsApiPath.ROOT,
+			preHandlers: [checkUserPermissions([PermissionKey.VIEW_ALL_PROJECTS])],
 			validation: {
 				body: projectCreateValidationSchema,
 			},
@@ -75,6 +76,7 @@ class ProjectController extends BaseController {
 				),
 			method: "DELETE",
 			path: ProjectsApiPath.$ID,
+			preHandlers: [checkUserPermissions([PermissionKey.VIEW_ALL_PROJECTS])],
 		});
 
 		this.addRoute({
@@ -111,6 +113,7 @@ class ProjectController extends BaseController {
 				),
 			method: "PATCH",
 			path: ProjectsApiPath.$ID,
+			preHandlers: [checkUserPermissions([PermissionKey.VIEW_ALL_PROJECTS])],
 			validation: {
 				body: projectPatchValidationSchema,
 			},
@@ -135,8 +138,8 @@ class ProjectController extends BaseController {
 	 *                description:
 	 *                  type: string
 	 *      responses:
-	 *        200:
-	 *          description: Successful operation
+	 *        201:
+	 *          description: Project was successfully created
 	 *          content:
 	 *            application/json:
 	 *              schema:
@@ -153,7 +156,7 @@ class ProjectController extends BaseController {
 	): Promise<APIHandlerResponse> {
 		return {
 			payload: await this.projectService.create(options.body),
-			status: HTTPCode.OK,
+			status: HTTPCode.CREATED,
 		};
 	}