+++ apps/frontend/src/libs/components/table/table.tsx
@@ -14,6 +14,8 @@ import styles from "./styles.module.css";
 type BaseProperties<T> = {
 	columns: TableColumn<T>[];
 	data: T[];
+	emptyPlaceholder?: string;
+	isScrollDisabled?: boolean;
 };
 
 type SelectableProperties<T> = {
@@ -29,6 +31,8 @@ type Properties<T> =
 const Table = <T extends object>({
 	columns,
 	data,
+	emptyPlaceholder = "There is nothing yet.",
+	isScrollDisabled,
 	...selectableProperties
 }: Properties<T>): JSX.Element => {
 	const { getRowId, onRowSelect, selectedRowIds } = selectableProperties as
@@ -46,7 +50,12 @@ const Table = <T extends object>({
 
 	return (
 		<div className={styles["table-container"]}>
-			<table className={styles["table"]}>
+			<table
+				className={getValidClassNames(
+					styles["table"],
+					isScrollDisabled && styles["table-no-scroll"],
+				)}
+			>
 				<thead className={styles["table-head"]}>
 					{table.getHeaderGroups().map((headerGroup) => (
 						<tr className={styles["table-row"]} key={headerGroup.id}>
@@ -62,9 +71,7 @@ const Table = <T extends object>({
 								<th
 									className={styles["table-header"]}
 									key={header.id}
-									style={{
-										width: header.column.columnDef.size,
-									}}
+									style={{ width: header.column.columnDef.size }}
 								>
 									{flexRender(
 										header.column.columnDef.header,
@@ -99,9 +106,7 @@ const Table = <T extends object>({
 									<td
 										className={styles["table-data"]}
 										key={cell.id}
-										style={{
-											width: cell.column.columnDef.size,
-										}}
+										style={{ width: cell.column.columnDef.size }}
 									>
 										{flexRender(cell.column.columnDef.cell, cell.getContext())}
 									</td>
@@ -112,7 +117,7 @@ const Table = <T extends object>({
 						<tr className={styles["table-row"]}>
 							<td className={styles["table-data"]} colSpan={columns.length}>
 								<p className={styles["empty-placeholder"]}>
-									There is nothing yet.
+									{emptyPlaceholder}
 								</p>
 							</td>
 						</tr>