+++ apps/frontend/src/pages/access-management/libs/components/group-users-table/libs/components/components.tsx
@@ -1,0 +1,1 @@
+export { GroupUsersSearch } from "./group-users-search/group-users-search.js";