+++ packages/shared/src/libs/helpers/helpers.ts
@@ -1,5 +1,8 @@
+export { addDays } from "./add-days/add-days.helper.js";
 export { changeCase } from "./change-case/change-case.helper.js";
 export { checkHasPermission } from "./check-has-permission/check-has-permission.helper.js";
 export { configureString } from "./configure-string/configure-string.helper.js";
 export { formatDate } from "./format-date/format-date.helper.js";
+export { getDifferenceInDays } from "./get-difference-in-days/get-difference-in-days.helper.js";
 export { initDebounce } from "./init-debounce/init-debounce.helper.js";
+export { subtractDays } from "./subtract-days/subtract-days.helper.js";