+++ packages/shared/src/modules/git-emails/libs/types/git-email-get-all-item-response-dto.type.ts
@@ -1,0 +1,10 @@
+type GitEmailGetAllItemResponseDto = {
+	contributor: {
+		id: number;
+		name: string;
+	};
+	email: string;
+	id: number;
+};
+
+export { type GitEmailGetAllItemResponseDto };