+++ apps/frontend/CHANGELOG.md
@@ -1,5 +1,46 @@
 # Changelog
 
+## [1.26.1](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/frontend-v1.26.0...frontend-v1.26.1) (2024-09-16)
+
+
+### Bug Fixes
+
+* Inappropriate appearance of the Access Managment page for a small screen resolution (1024px) gf-295 ([#323](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/323)) ([0588959](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/058895952437ddff369edc7d14e30066a2f87df6))
+* Upon a click on the Logo element in the header user is redirected to the Homepage gf-285 ([#291](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/291)) ([fc2ba06](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/fc2ba067490675837e6f285aebdf40cff2009320))
+
+## [1.26.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/frontend-v1.25.0...frontend-v1.26.0) (2024-09-16)
+
+
+### Features
+
+* add an ability to regenerate project api key gf-288 ([#307](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/307)) ([7551326](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/75513264ccf452c8bd1c9f9ff99bf90f71b040f7))
+
+
+### Bug Fixes
+
+* adjusted project status message based on search state gf-313 ([#319](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/319)) ([0306fa4](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/0306fa4879cfa077cb53e9dd067e62ed0e8fd08a))
+* The Table should not be able to scroll when the Options Popover is open gf-272 ([#278](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/278)) ([8450514](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/8450514f1fcd9f913cf8bd6716c88a975cb34e69))
+
+## [1.25.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/frontend-v1.24.0...frontend-v1.25.0) (2024-09-16)
+
+
+### Features
+
+* add date input to analytics page gf-182 ([#237](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/237)) ([9570a5b](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/9570a5b5b9f34bcd82bc5a4cbcfe365f127b9c8b))
+* add users search in group forms gf-270 ([#303](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/303)) ([985d160](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/985d160827edf30b59c4ea38cf09813015303068))
+
+## [1.24.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/frontend-v1.23.0...frontend-v1.24.0) (2024-09-13)
+
+
+### Features
+
+* Add an ability to copy project api key gf-284 ([#292](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/292)) ([1251eb9](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/1251eb969f3f6dd8f1abe7eb7887605f2931d410))
+
+
+### Bug Fixes
+
+* rename setup analytics button to match the design gf-308 ([#309](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/issues/309)) ([abcbde2](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/commit/abcbde2612c4c40215f188b748408db42ecd2835))
+
 ## [1.23.0](https://github.com/BinaryStudioAcademy/bsa-2024-gitfit/compare/frontend-v1.22.1...frontend-v1.23.0) (2024-09-13)